Move common intl strings
