+++ src/app/intl/IntlProvider.tsx
@@ -8,6 +8,8 @@ import nnMessages from './nn_NO.json';
 import nbMessages from './nb_NO.json';
 import nnMessagesUttaksplan from '../../uttaksplan/intl/nn_NO.json';
 import nbMessagesUttaksplan from '../../uttaksplan/intl/nb_NO.json';
+import nnMessagesCommon from '../../common/intl/nn_NO.json';
+import nbMessagesCommon from '../../common/intl/nb_NO.json';
 import { AppState } from '../redux/reducers';
 import { Språkkode } from 'common/intl/types';
 
@@ -24,8 +26,16 @@ class IntlProvider extends React.Component<StateProps> {
     render() {
         const messages =
             this.props.språkkode === 'nb'
-                ? { ...nbMessages, ...nbMessagesUttaksplan }
-                : { ...nnMessages, ...nnMessagesUttaksplan };
+                ? {
+                      ...nbMessages,
+                      ...nbMessagesUttaksplan,
+                      ...nbMessagesCommon
+                  }
+                : {
+                      ...nnMessages,
+                      ...nnMessagesUttaksplan,
+                      ...nnMessagesCommon
+                  };
         return (
             <Provider locale="nb" messages={messages || {}}>
                 {this.props.children}
