+++ src/common/intl/nb_NO.json
@@ -1,0 +1,17 @@
+{
+    "common.varighet.dager": "{dager, plural, one {# dag} other {# dager}}",
+    "common.varighet.uker": "{uker, plural, one {# uke} other {# uker}}",
+    "common.varighet.ukerogdager":
+        "{uker, plural, zero {} one {# uke} other {# uker}} {dager, plural, =0 {} one {og # dag} other {og # dager}}",
+    "common.varighet.og": "og",
+    "common.tall-1": "én",
+    "common.tall-2": "to",
+    "common.tall-3": "tre",
+    "common.tall-4": "fire",
+    "common.tall-5": "fem",
+    "common.tall-6": "seks",
+    "common.tall-7": "syv",
+    "common.tall-8": "åtte",
+    "common.tall-9": "ni",
+    "common.tall-10": "ti"
+}
