+++ src/common/intl/nn_NO.json
@@ -1,0 +1,17 @@
+{
+    "common.varighet.dager": "{dager, plural, one {# dag} other {# dagar}}",
+    "common.varighet.uker": "{uker, plural, one {# veke} other {# veker}}",
+    "common.varighet.ukerogdager":
+        "{uker, plural, zero {} one {# veke} other {# veker}} {dager, plural, =0 {} one {og # dag} other {og # dagar}}",
+    "common.tall-1": "éin",
+    "common.tall-2": "to",
+    "common.tall-3": "tre",
+    "common.tall-4": "fire",
+    "common.tall-5": "fem",
+    "common.tall-6": "seks",
+    "common.tall-7": "sju",
+    "common.tall-8": "åtte",
+    "common.tall-9": "ni",
+    "common.tall-10": "ti",
+    "common.og": "og"
+}
