+++ src/common/util/datoUtils.ts
@@ -30,3 +30,13 @@ export function ukedagKort(dato: Date): string {
 export function dagIMåned(dato: Date): string {
     return format(dato, 'D.', { locale });
 }
+
+export const getUkerOgDagerFromDager = (
+    dager: number
+): { uker: number; dager: number } => {
+    const uker = Math.floor(dager / 5);
+    return {
+        dager: dager - uker * 5,
+        uker
+    };
+};
