+++ src/uttaksplan/intl/nb_NO.json
@@ -4,21 +4,6 @@
     "uttaksplan.Forelder2": "Far eller medmor",
     "uttaksplan.forelder1": "mor",
     "uttaksplan.forelder2": "far eller medmor",
-    "uttaksplan.dager": "{dager, plural, one {# dag} other {# dager}}",
-    "uttaksplan.uker": "{uker, plural, one {# uke} other {# uker}}",
-    "uttaksplan.ukerogdager":
-        "{uker, plural, zero {} one {# uke} other {# uker}} {dager, plural, =0 {} one {og # dag} other {og # dager}}",
-    "uttaksplan.tall-1": "én",
-    "uttaksplan.tall-2": "to",
-    "uttaksplan.tall-3": "tre",
-    "uttaksplan.tall-4": "fire",
-    "uttaksplan.tall-5": "fem",
-    "uttaksplan.tall-6": "seks",
-    "uttaksplan.tall-7": "syv",
-    "uttaksplan.tall-8": "åtte",
-    "uttaksplan.tall-9": "ni",
-    "uttaksplan.tall-10": "ti",
-    "uttaksplan.og": "og",
     "uttaksplan.utvidetinfo.lesmer": "Les mer",
     "uttaksplan.utvidetinfo.lukk": "Lukk",
     "uttaksplan.startdato": "Startdato",
