+++ src/uttaksplan/intl/nn_NO.json
@@ -4,21 +4,6 @@
     "uttaksplan.Forelder2": "Far eller medmor",
     "uttaksplan.forelder1": "mor",
     "uttaksplan.forelder2": "far eller medmor",
-    "uttaksplan.dager": "{dager, plural, one {# dag} other {# dagar}}",
-    "uttaksplan.uker": "{uker, plural, one {# veke} other {# veker}}",
-    "uttaksplan.ukerogdager":
-        "{uker, plural, zero {} one {# veke} other {# veker}} {dager, plural, =0 {} one {og # dag} other {og # dagar}}",
-    "uttaksplan.tall-1": "éin",
-    "uttaksplan.tall-2": "to",
-    "uttaksplan.tall-3": "tre",
-    "uttaksplan.tall-4": "fire",
-    "uttaksplan.tall-5": "fem",
-    "uttaksplan.tall-6": "seks",
-    "uttaksplan.tall-7": "sju",
-    "uttaksplan.tall-8": "åtte",
-    "uttaksplan.tall-9": "ni",
-    "uttaksplan.tall-10": "ti",
-    "uttaksplan.og": "og",
     "uttaksplan.utvidetinfo.lesmer": "Les meir",
     "uttaksplan.utvidetinfo.lukk": "Lukk",
     "uttaksplan.startdato": "Startdato",
