+++ src/uttaksplan/utils/uttaksdagerUtils.ts
@@ -98,15 +98,6 @@ export interface UkerOgDager {
     dager: number;
     uker: number;
 }
-export const getUkerOgDagerFromDager = (
-    dager: number
-): { uker: number; dager: number } => {
-    const uker = Math.floor(dager / 5);
-    return {
-        dager: dager - uker * 5,
-        uker
-    };
-};
 
 /**
  * Legger til dager til en dato og returnerer ny dato
